requirements.txt
Alfred-Workflow==1.40.0
gkeepapi>=0.10.6